use modernized next structure
