+++ src/app/about/page.tsx
@@ -1,0 +1,22 @@
+//import NestedLayout from '../../components/nested-layout'
+// import type { NextPageWithLayout } from './_app.tsx'
+
+// const Page: NextPageWithLayout = () => {
+//     return <p>hello world</p>
+//   }
+   
+//   Page.getLayout = function getLayout(page: ReactElement) {
+//     return (
+//       <Layout>
+//         <NestedLayout>{page}</NestedLayout>
+//       </Layout>
+//     )
+//   }
+   
+//   export default Page
+
+function About() {
+    return <div>Example About</div>
+  }
+
+export default About;
