+++ src/app/home/page.tsx
@@ -1,0 +1,14 @@
+'use client'
+ 
+// This is a Client Component (same as components in the `pages` directory)
+// It receives data as props, has access to state and effects, and is
+// prerendered on the server during the initial page load.
+function Home() {
+  return (
+    <div>
+      Example home... page?
+    </div>
+  )
+}
+
+export default Home;
