+++ src/components/footer/index.tsx
@@ -1,0 +1,6 @@
+function Footer() {
+    // TODO: add navbar items
+      return <div>sample footer</div>
+  }
+  
+  export default Footer;
