+++ src/components/header/index.tsx
@@ -1,0 +1,6 @@
+function Header() {
+  // TODO: add navbar items
+    return <div>sample header</div>
+}
+
+export default Header;
